Get the horse breeding/drinking updates to run
FeedableInventorySystem_Update seems to be deprecated.
We don't really need to patch it though... we just need our own system to run.
So it now hooks CheckInSunSystem instead, just because this system seems to always run even during nighttime.
There's probably a better way to do it... but this seems to work just fine
